refactor(core)!: move `metadata::Metadata` to `ast::Metadata`
