+++ conjure_oxide/src/lib.rs
@@ -2,8 +2,8 @@
 
 pub use conjure_cp_core::Model;
 pub use conjure_cp_core::ast;
+pub use conjure_cp_core::ast::Metadata;
 pub use conjure_cp_core::error::Error;
-pub use conjure_cp_core::metadata::Metadata;
 pub use conjure_cp_core::parse::{get_example_model, get_example_model_by_path, model_from_json};
 pub use conjure_cp_core::rule_engine;
 pub use conjure_cp_core::rule_engine::{
