+++ crates/conjure-cp-core/src/lib.rs
@@ -12,7 +12,6 @@ mod bug;
 
 pub mod context;
 pub mod error;
-pub mod metadata;
 pub mod parse;
 pub mod representation;
 pub mod rule_engine;
