+++ crates/conjure-cp-essence-parser/src/parser/expression.rs
@@ -1,8 +1,8 @@
 #![allow(clippy::legacy_numeric_constants)]
 use tree_sitter::Node;
 
+use conjure_cp_core::ast::Metadata;
 use conjure_cp_core::ast::{Atom, Expression, Literal, Moo, Name, SymbolTable};
-use conjure_cp_core::metadata::Metadata;
 use conjure_cp_core::{into_matrix_expr, matrix_expr};
 
 use crate::errors::EssenceParseError;
