+++ crates/conjure-cp-rules/src/matrix/repr_matrix.rs
@@ -1,9 +1,9 @@
+use conjure_cp_core::ast::Metadata;
 use conjure_cp_core::ast::categories::{Category, CategoryOf};
 use conjure_cp_core::ast::{
     Atom, Domain, Expression as Expr, Literal, Moo, Name, Range, SymbolTable, matrix,
 };
 use conjure_cp_core::into_matrix_expr;
-use conjure_cp_core::metadata::Metadata;
 use conjure_cp_core::rule_engine::{
     ApplicationError::RuleNotApplicable, ApplicationResult, Reduction, register_rule,
 };
