+++ crates/conjure-cp-rules/src/sets/horizontal/concat.rs
@@ -1,8 +1,8 @@
 // rules for concatenations of subsetEq with intersect and union
 // analogous rules for subset, supset and supsetEq are not needed because these are converted to subsetEq first.
+use conjure_cp_core::ast::Metadata;
 use conjure_cp_core::ast::{Expression as Expr, Moo, ReturnType, SymbolTable, Typeable};
 use conjure_cp_core::matrix_expr;
-use conjure_cp_core::metadata::Metadata;
 use conjure_cp_core::rule_engine::Reduction;
 use conjure_cp_core::rule_engine::{
     ApplicationError::RuleNotApplicable, ApplicationResult, register_rule,
