+++ crates/conjure-cp-rules/src/sets/horizontal/equals.rs
@@ -1,8 +1,8 @@
 use conjure_cp_core::ast::Moo;
 // Equals rule for sets
+use conjure_cp_core::ast::Metadata;
 use conjure_cp_core::ast::{Expression, ReturnType::Set, SymbolTable, Typeable};
 use conjure_cp_core::matrix_expr;
-use conjure_cp_core::metadata::Metadata;
 use conjure_cp_core::rule_engine::Reduction;
 use conjure_cp_core::rule_engine::{
     ApplicationError::RuleNotApplicable, ApplicationResult, register_rule,
