+++ crates/conjure-cp-rules/src/sets/horizontal/supseteq.rs
@@ -1,6 +1,6 @@
 // SupsetEq rule for sets
+use conjure_cp_core::ast::Metadata;
 use conjure_cp_core::ast::{Expression as Expr, ReturnType, SymbolTable, Typeable};
-use conjure_cp_core::metadata::Metadata;
 use conjure_cp_core::rule_engine::Reduction;
 use conjure_cp_core::rule_engine::{
     ApplicationError::RuleNotApplicable, ApplicationResult, register_rule,
