+++ crates/conjure-cp-rules/src/utils.rs
@@ -1,6 +1,6 @@
 use conjure_cp_core::{
+    ast::Metadata,
     ast::{Atom, DeclarationPtr, Expression as Expr, Moo, SymbolTable, categories::Category},
-    metadata::Metadata,
 };
 
 use tracing::{instrument, trace};
